requirements.txt
Django==3.0
django-allauth==0.54.0